explorations/ai-responsibility-index/requirements.txt
pymupdf>=1.23.0
lxml>=4.9.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
anthropic>=0.18.0
numpy>=1.24.0
